Move/rename/adjust `prepare_behaviour_dump_dir`
The destination is `crate::core::builder::cargo`, the new name is
`prepare_shims_dump_dir`, and the adjustment is to take `&Builder<'_>` instead
of `&Build`.
